INT-67 - Introduce HasBeenProcessed property
Introduced a new property "HasBeenProcessed" on IMarketState. This
property is set to true only when the market is passed to the plugin to
be processed (doesn't mean that the market is created on the downstream
system though)

It allows other rules to take proper action is the market has or has not
been seen by the plugin.